+++ HISTORY.md
@@ -70,6 +70,7 @@
     * fixed zebra coloring when importing a script from file or smart PNG
     * fixed accidentally triggering "When I receive green flag" events for any broadcast message
     * COMBINing an empty list with JOIN results in an empty text rather than in the number zero
+    * hide new primitives in old microworlds
 * **Documentation Updates:**
     * new "docs" folder for documentation files, thanks, Michael!
     * new doc for extensions, thanks, Michael!
@@ -77,6 +78,9 @@
     * German
     * Catalan, thanks, Joan!
     * Dutch, thanks, Hans!
+
+### 2023-10-09
+* store, objects: hide new primitives in old microworlds
 
 ### 2023-10-05
 * threads: COMBINing an empty list with JOIN results in an empty text rather than in the number zero